Widen Fill length parameter from byte to integer
Fill(len: byte; ...) silently truncated len to 8 bits when given a
word-typed argument. snviewer.pas:249 calls Fill(ScrWidth-2, '═') to
draw the viewer's top border. On a terminal wider than 257 columns,
ScrWidth-2 wrapped through byte and produced a foreshortened border;
on a 300-col terminal the border was 42 chars instead of 298.

sn_obj.pas's local Fill already used integer; aligning RV's Fill
(in rv.pas / rvstr.inc) and trd_ovr.pas's local fill matches that
choice. Each definition now guards len <= 0 since DupeString rejects
negative counts.

+++ lib/rv/rv.pas
@@ -255,7 +255,7 @@ function ConsoleText(const s: string): string;
 function NoSpace(s: string): string;
 function NoSpaceLR(s: string): string;
 function Vall(tempein: string): longint;
-function Fill(len: byte; const symb: string): string;
+function Fill(len: integer; const symb: string): string;
 function WithOut(s, chars: string): string;
 function CLen(s: string): integer;
 function LZ(w: word): string;
